requirements.txt
discord.py
requests
python-dotenv
pyTelegramBotAPI
